refactor: Split codecs into separate files, keep Tim's style
Move each codec class out of ClpConnectorCodecProvider into its own
file. ClpConnectorCodecProvider now only provides the connector-level
factory methods. Each codec still follows Tim's style from #26650
(cast to type-specific variable, UncheckedIOException, plain try).
